+++ common/src/main/java/io/github/sds100/keymapper/common/utils/NodeInteractionType.kt
@@ -2,6 +2,7 @@ package io.github.sds100.keymapper.common.utils
 
 import android.view.accessibility.AccessibilityNodeInfo
 
+// DO NOT change these values, they are used in the data schema.
 enum class NodeInteractionType(val accessibilityActionId: Int) {
     CLICK(AccessibilityNodeInfo.ACTION_CLICK),
     LONG_CLICK(AccessibilityNodeInfo.ACTION_LONG_CLICK),